+++ packages/specifications/src/raccordement/stepDefinitions/propositionTechniqueEtFinancière.given.ts
@@ -1,0 +1,30 @@
+import { Given as EtantDonné } from '@cucumber/cucumber';
+import { mediator } from 'mediateur';
+
+import { Raccordement } from '@potentiel-domain/reseau';
+
+import { PotentielWorld } from '../../potentiel.world';
+
+EtantDonné(
+  'une proposition technique et financière pour le dossier de raccordement du projet lauréat',
+  async function (this: PotentielWorld) {
+    const { identifiantProjet, référenceDossier } =
+      this.raccordementWorld.transmettreDemandeComplèteRaccordementFixture;
+
+    const { dateSignature, propositionTechniqueEtFinancièreSignée } =
+      this.raccordementWorld.transmettrePropositionTechniqueEtFinancièreFixture.créer({
+        identifiantProjet,
+        référenceDossier,
+      });
+
+    await mediator.send<Raccordement.RaccordementUseCase>({
+      type: 'Réseau.Raccordement.UseCase.TransmettrePropositionTechniqueEtFinancière',
+      data: {
+        identifiantProjetValue: identifiantProjet,
+        référenceDossierRaccordementValue: référenceDossier,
+        dateSignatureValue: dateSignature,
+        propositionTechniqueEtFinancièreSignéeValue: propositionTechniqueEtFinancièreSignée,
+      },
+    });
+  },
+);
