+++ packages/specifications/src/raccordement/stepDefinitions/raccordement.given.ts
@@ -1,16 +1,13 @@
-import { DataTable, Given as EtantDonné } from '@cucumber/cucumber';
+import { Given as EtantDonné } from '@cucumber/cucumber';
 import { mediator } from 'mediateur';
 import waitForExpect from 'wait-for-expect';
 import { expect } from 'chai';
 
-import { DateTime } from '@potentiel-domain/common';
 import { GestionnaireRéseau, Raccordement } from '@potentiel-domain/reseau';
 import { ListerTâchesQuery } from '@potentiel-domain/tache';
 
 import { PotentielWorld } from '../../potentiel.world';
 import { RechercherTypeTâche } from '../../tâche/tâche.world';
-
-import { transmettreDemandeComplèteRaccordement } from './raccordement.when';
 
 EtantDonné(
   'le gestionnaire de réseau {string} attribué au raccordement du projet lauréat',
@@ -44,75 +41,6 @@ EtantDonné(
         rôleValue: this.utilisateurWorld.adminFixture.role,
       },
     });
-  },
-);
-
-EtantDonné(
-  'une demande complète de raccordement pour le projet lauréat',
-  async function (this: PotentielWorld) {
-    await transmettreDemandeComplèteRaccordement.call(this, this.lauréatWorld.identifiantProjet);
-  },
-);
-
-EtantDonné(
-  'une demande complète de raccordement pour le projet lauréat avec :',
-  async function (this: PotentielWorld, datatable: DataTable) {
-    await transmettreDemandeComplèteRaccordement.call(
-      this,
-      this.lauréatWorld.identifiantProjet,
-      datatable.rowsHash(),
-    );
-  },
-);
-
-EtantDonné(
-  'une proposition technique et financière pour le dossier de raccordement du projet lauréat',
-  async function (this: PotentielWorld) {
-    const { identifiantProjet, référenceDossier } =
-      this.raccordementWorld.transmettreDemandeComplèteRaccordementFixture;
-
-    const { dateSignature, propositionTechniqueEtFinancièreSignée } =
-      this.raccordementWorld.transmettrePropositionTechniqueEtFinancièreFixture.créer({
-        identifiantProjet,
-        référenceDossier,
-      });
-
-    await mediator.send<Raccordement.RaccordementUseCase>({
-      type: 'Réseau.Raccordement.UseCase.TransmettrePropositionTechniqueEtFinancière',
-      data: {
-        identifiantProjetValue: identifiantProjet,
-        référenceDossierRaccordementValue: référenceDossier,
-        dateSignatureValue: dateSignature,
-        propositionTechniqueEtFinancièreSignéeValue: propositionTechniqueEtFinancièreSignée,
-      },
-    });
-  },
-);
-
-EtantDonné(
-  'une date de mise en service pour le dossier de raccordement du projet lauréat',
-  async function (this: PotentielWorld) {
-    const { identifiantProjet } = this.lauréatWorld;
-    const { référenceDossier } = this.raccordementWorld;
-    const { dateMiseEnService } = this.raccordementWorld.transmettreDateMiseEnServiceFixture.créer({
-      identifiantProjet: identifiantProjet.formatter(),
-      référenceDossier,
-    });
-
-    try {
-      await mediator.send<Raccordement.RaccordementUseCase>({
-        type: 'Réseau.Raccordement.UseCase.TransmettreDateMiseEnService',
-        data: {
-          identifiantProjetValue: identifiantProjet.formatter(),
-          référenceDossierValue: référenceDossier,
-          dateMiseEnServiceValue: dateMiseEnService,
-          transmiseLeValue: DateTime.now().formatter(),
-          transmiseParValue: this.utilisateurWorld.adminFixture.email,
-        },
-      });
-    } catch (e) {
-      this.error = e as Error;
-    }
   },
 );
