+++ packages/specifications/src/raccordement/stepDefinitions/raccordement.then.ts
@@ -7,10 +7,8 @@ import { Raccordement, GestionnaireRéseau } from '@potentiel-domain/reseau';
 import { Option } from '@potentiel-libraries/monads';
 import { IdentifiantProjet } from '@potentiel-domain/common';
 import { mapToPlainObject } from '@potentiel-domain/core';
-import { ConsulterDocumentProjetQuery } from '@potentiel-domain/document';
 
 import { PotentielWorld } from '../../potentiel.world';
-import { convertReadableStreamToString } from '../../helpers/convertReadableToString';
 
 Alors(
   `le dossier est consultable dans la liste des dossiers de raccordement du projet lauréat`,
@@ -31,57 +29,6 @@ Alors(
       actual.dossiers
         .map((d) => d.référence.formatter())
         .should.contain(this.raccordementWorld.référenceDossier);
-    });
-  },
-);
-
-Alors(
-  `la demande complète de raccordement devrait être consultable dans le dossier de raccordement du projet lauréat`,
-  async function (this: PotentielWorld) {
-    const { identifiantProjet } = this.lauréatWorld;
-    const { référenceDossier } = this.raccordementWorld;
-    await waitForExpect(async () => {
-      const raccordement = await mediator.send<Raccordement.ConsulterRaccordementQuery>({
-        type: 'Réseau.Raccordement.Query.ConsulterRaccordement',
-        data: {
-          identifiantProjetValue: identifiantProjet.formatter(),
-        },
-      });
-
-      const dossierRaccordement =
-        await mediator.send<Raccordement.ConsulterDossierRaccordementQuery>({
-          type: 'Réseau.Raccordement.Query.ConsulterDossierRaccordement',
-          data: {
-            identifiantProjetValue: identifiantProjet.formatter(),
-            référenceDossierRaccordementValue: référenceDossier,
-          },
-        });
-
-      vérifierRaccordement.call(this, identifiantProjet, raccordement);
-      vérifierDossierRaccordement.call(this, identifiantProjet, dossierRaccordement);
-
-      assert(Option.isSome(dossierRaccordement), 'dossierRaccordement is undefined');
-      assert(
-        dossierRaccordement.demandeComplèteRaccordement,
-        'demandeComplèteRaccordement is undefined',
-      );
-      const actualAccuséRéception = dossierRaccordement.demandeComplèteRaccordement.accuséRéception;
-      const { accuséRéception: expectedAccuséRéception } = this.raccordementWorld
-        .modifierDemandeComplèteRaccordementFixture.aÉtéCréé
-        ? this.raccordementWorld.modifierDemandeComplèteRaccordementFixture
-        : this.raccordementWorld.transmettreDemandeComplèteRaccordementFixture;
-
-      assert(actualAccuséRéception, 'actualAccuséRéception is not defined');
-      const result = await mediator.send<ConsulterDocumentProjetQuery>({
-        type: 'Document.Query.ConsulterDocumentProjet',
-        data: {
-          documentKey: actualAccuséRéception.formatter(),
-        },
-      });
-
-      const actualContent = await convertReadableStreamToString(result.content);
-      const expectedContent = await convertReadableStreamToString(expectedAccuséRéception.content);
-      actualContent.should.be.equal(expectedContent);
     });
   },
 );
@@ -149,71 +96,6 @@ Alors(
       }
 
       actual.dossiers.should.length(nombreDeDemandes);
-    });
-  },
-);
-
-Alors(
-  `la date de mise en service devrait être consultable dans le dossier de raccordement du projet lauréat`,
-  async function (this: PotentielWorld) {
-    const { identifiantProjet } = this.lauréatWorld;
-    const { référenceDossier } = this.raccordementWorld;
-    await waitForExpect(async () => {
-      const dossierRaccordement =
-        await mediator.send<Raccordement.ConsulterDossierRaccordementQuery>({
-          type: 'Réseau.Raccordement.Query.ConsulterDossierRaccordement',
-          data: {
-            référenceDossierRaccordementValue: référenceDossier,
-            identifiantProjetValue: identifiantProjet.formatter(),
-          },
-        });
-
-      vérifierDossierRaccordement.call(this, identifiantProjet, dossierRaccordement);
-    });
-  },
-);
-
-Alors(
-  `la proposition technique et financière signée devrait être consultable dans le dossier de raccordement du projet lauréat`,
-  async function (this: PotentielWorld) {
-    const { identifiantProjet } = this.lauréatWorld;
-    const { référenceDossier } = this.raccordementWorld;
-    await waitForExpect(async () => {
-      const dossierRaccordement =
-        await mediator.send<Raccordement.ConsulterDossierRaccordementQuery>({
-          type: 'Réseau.Raccordement.Query.ConsulterDossierRaccordement',
-          data: {
-            identifiantProjetValue: identifiantProjet.formatter(),
-            référenceDossierRaccordementValue: référenceDossier,
-          },
-        });
-
-      vérifierDossierRaccordement.call(this, identifiantProjet, dossierRaccordement);
-      assert(Option.isSome(dossierRaccordement));
-
-      const { propositionTechniqueEtFinancière } = dossierRaccordement;
-
-      assert(propositionTechniqueEtFinancière, 'propositionTechniqueEtFinancière is undefined');
-
-      const { propositionTechniqueEtFinancièreSignée } = this.raccordementWorld
-        .modifierPropositionTechniqueEtFinancièreFixture.aÉtéCréé
-        ? this.raccordementWorld.modifierPropositionTechniqueEtFinancièreFixture
-        : this.raccordementWorld.transmettrePropositionTechniqueEtFinancièreFixture;
-
-      const result = await mediator.send<ConsulterDocumentProjetQuery>({
-        type: 'Document.Query.ConsulterDocumentProjet',
-        data: {
-          documentKey:
-            propositionTechniqueEtFinancière.propositionTechniqueEtFinancièreSignée.formatter(),
-        },
-      });
-
-      const actualContent = await convertReadableStreamToString(result.content);
-      const expectedContent = await convertReadableStreamToString(
-        propositionTechniqueEtFinancièreSignée.content,
-      );
-
-      actualContent.should.be.equal(expectedContent);
     });
   },
 );
@@ -323,32 +205,7 @@ async function vérifierGestionnaireAttribué(
   });
 }
 
-function vérifierRaccordement(
-  this: PotentielWorld,
-  identifiantProjet: IdentifiantProjet.ValueType,
-  raccordement: Option.Type<Raccordement.ConsulterRaccordementReadModel>,
-) {
-  const { raccordement: expectedRaccordement } = mapToPlainObject(
-    this.raccordementWorld.mapToExpected(identifiantProjet),
-  );
-  const actualRaccordement = mapToPlainObject(raccordement);
-
-  // HACK : misÀJourLe vaut la date de l'event (created_at),
-  //  on ne peut pas calculer cette date de manière exacte dans les tests
-  if (Option.isSome(actualRaccordement)) {
-    for (const doss of actualRaccordement.dossiers) {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      delete (doss as any).misÀJourLe;
-    }
-  }
-
-  actualRaccordement.should.be.deep.equal(
-    expectedRaccordement,
-    `le raccordement n'est pas identique`,
-  );
-}
-
-function vérifierDossierRaccordement(
+export function vérifierDossierRaccordement(
   this: PotentielWorld,
   identifiantProjet: IdentifiantProjet.ValueType,
   dossierRaccordement: Option.Type<Raccordement.ConsulterDossierRaccordementReadModel>,
